Удалена камера для 3Д, камера для 2Д переименована
