+++ src/main/java/amaralus/apps/hackandslash/graphics/SpriteRenderer.java
@@ -1,6 +1,6 @@
 package amaralus.apps.hackandslash.graphics;
 
-import amaralus.apps.hackandslash.graphics.camera.OrthoCamera;
+import amaralus.apps.hackandslash.graphics.camera.Camera;
 import amaralus.apps.hackandslash.graphics.data.sprites.Sprite;
 import org.joml.Matrix4f;
 import org.joml.Vector2f;
@@ -17,7 +17,7 @@ public class SpriteRenderer {
         textureShader = new Shader("vertex", "fragment");
     }
 
-    public void draw(OrthoCamera camera, Sprite sprite, Vector2f entityPos, float rotateAngle) {
+    public void draw(Camera camera, Sprite sprite, Vector2f entityPos, float rotateAngle) {
         var textureSize = camera.getSpriteScaleOfCam(sprite);
         var cameraEntityPos = camera.getEntityCamPos(entityPos, textureSize, sprite.getOffsetToSpriteCenter());
